requirements.txt
flask==2.3.3
flask-mysqldb==1.0.1
flask-session==0.5.0
flask-login==0.6.3
python-dotenv==1.0.0
Werkzeug==2.3.7
mysql-connector-python==8.2.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
gunicorn==21.2.0
redis==5.0.1 